coop-langchain-full/requirements.txt
fastapi>=0.112
uvicorn[standard]>=0.30
requests>=2.32
python-dotenv>=1.0
orjson>=3.10
PyYAML>=6.0
gspread>=6.1
google-auth>=2.29
google-api-python-client>=2.136
upstash-redis>=1.1.0

# LangChain + Gemini
langchain>=0.2.12
langchain-google-genai>=2.0.0
pydantic>=2.7
